Moved: Entity/*.php -> Model/*.php
