+++ README.md
@@ -85,14 +85,14 @@ loevgaard_sylius_brand:
 
 ```php
 <?php
-// src/Entity/Product.php
+// src/Model/Product.php
 
 declare(strict_types=1);
 
-namespace App\Entity;
+namespace App\Model;
 
-use Loevgaard\SyliusBrandPlugin\Entity\ProductTrait;
-use Loevgaard\SyliusBrandPlugin\Entity\ProductInterface as LoevgaardSyliusBrandPluginProductInterface;
+use Loevgaard\SyliusBrandPlugin\Model\ProductTrait;
+use Loevgaard\SyliusBrandPlugin\Model\ProductInterface as LoevgaardSyliusBrandPluginProductInterface;
 use Sylius\Component\Core\Model\Product as BaseProduct;
 
 class Product extends BaseProduct implements LoevgaardSyliusBrandPluginProductInterface
@@ -106,15 +106,15 @@ class Product extends BaseProduct implements LoevgaardSyliusBrandPluginProductIn
 ```xml
 <?xml version="1.0" encoding="UTF-8"?>
 
-<!-- config/doctrine/Product.orm.xml -->
+<!-- config/doctrine/model/Product.orm.xml -->
 
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                                       http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <mapped-superclass name="App\Entity\Product" table="sylius_product">
-        <many-to-one field="brand" target-entity="Loevgaard\SyliusBrandPlugin\Entity\Brand">
+    <mapped-superclass name="App\Model\Product" table="sylius_product">
+        <many-to-one field="brand" target-entity="Loevgaard\SyliusBrandPlugin\Model\Brand" inversed-by="products">
             <join-column name="brand_id" on-delete="SET NULL" />
         </many-to-one>
     </mapped-superclass>
@@ -155,7 +155,7 @@ sylius_product:
     resources:
         product:
             classes:
-                model: App\Entity\Product
+                model: App\Model\Product
                 repository: App\Doctrine\ORM\ProductRepository
 
 ```
