+++ src/Doctrine/ORM/BrandRepositoryInterface.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Loevgaard\SyliusBrandPlugin\Doctrine\ORM;
 
 use Doctrine\ORM\QueryBuilder;
-use Loevgaard\SyliusBrandPlugin\Entity\BrandInterface;
+use Loevgaard\SyliusBrandPlugin\Model\BrandInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 
 interface BrandRepositoryInterface extends RepositoryInterface
