+++ src/Doctrine/ORM/ProductRepositoryInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Loevgaard\SyliusBrandPlugin\Doctrine\ORM;
 
-use Loevgaard\SyliusBrandPlugin\Entity\BrandInterface;
+use Loevgaard\SyliusBrandPlugin\Model\BrandInterface;
 use Sylius\Component\Core\Repository\ProductRepositoryInterface as BaseProductRepositoryInterface;
 
 interface ProductRepositoryInterface extends BaseProductRepositoryInterface
