+++ src/Factory/BrandImageFactory.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Loevgaard\SyliusBrandPlugin\Factory;
 
-use Loevgaard\SyliusBrandPlugin\Entity\BrandImageInterface;
-use Loevgaard\SyliusBrandPlugin\Entity\BrandInterface;
+use Loevgaard\SyliusBrandPlugin\Model\BrandImageInterface;
+use Loevgaard\SyliusBrandPlugin\Model\BrandInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 
 class BrandImageFactory implements BrandImageFactoryInterface
