+++ src/LoevgaardSyliusBrandPlugin.php
@@ -7,10 +7,10 @@ declare(strict_types=1);
 namespace Loevgaard\SyliusBrandPlugin;
 
 use Sylius\Bundle\CoreBundle\Application\SyliusPluginTrait;
+use Sylius\Bundle\ResourceBundle\AbstractResourceBundle;
 use Sylius\Bundle\ResourceBundle\SyliusResourceBundle;
-use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-final class LoevgaardSyliusBrandPlugin extends Bundle
+final class LoevgaardSyliusBrandPlugin extends AbstractResourceBundle
 {
     use SyliusPluginTrait;
