+++ src/Resources/config/doctrine/model/Brand.orm.xml
@@ -6,11 +6,11 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                                       http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <mapped-superclass name="Loevgaard\SyliusBrandPlugin\Entity\Brand" table="loevgaard_brand">
+    <mapped-superclass name="Loevgaard\SyliusBrandPlugin\Model\Brand" table="loevgaard_brand">
         <id name="id" column="id" type="integer">
             <generator strategy="AUTO" />
         </id>
-        <one-to-many field="images" target-entity="Loevgaard\SyliusBrandPlugin\Entity\BrandImage" mapped-by="owner" orphan-removal="true">
+        <one-to-many field="images" target-entity="Loevgaard\SyliusBrandPlugin\Model\BrandImage" mapped-by="owner" orphan-removal="true">
             <cascade>
                 <cascade-all/>
             </cascade>
