+++ src/Resources/config/doctrine/model/BrandImage.orm.xml
@@ -1,13 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<!--suppress XmlDefaultAttributeValue -->
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                                       http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <mapped-superclass name="Tests\Loevgaard\SyliusBrandPlugin\Application\Entity\Product" table="sylius_product">
-        <many-to-one field="brand" target-entity="Loevgaard\SyliusBrandPlugin\Entity\Brand">
-            <join-column name="brand_id" on-delete="SET NULL" />
+    <mapped-superclass name="Loevgaard\SyliusBrandPlugin\Model\BrandImage" table="loevgaard_brand_image">
+        <many-to-one field="owner" target-entity="Loevgaard\SyliusBrandPlugin\Model\Brand" inversed-by="images">
+            <join-column name="owner_id" referenced-column-name="id" nullable="false" on-delete="CASCADE" />
         </many-to-one>
     </mapped-superclass>
