+++ src/Resources/config/validation/Brand.xml
@@ -4,7 +4,7 @@
                     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                     xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
 
-    <class name="Loevgaard\SyliusBrandPlugin\Entity\Brand">
+    <class name="Loevgaard\SyliusBrandPlugin\Model\Brand">
 
         <constraint name="Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity">
             <option name="fields">slug</option>
