+++ src/Resources/config/validation/BrandImage.xml
@@ -4,7 +4,7 @@
                     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                     xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
 
-    <class name="Loevgaard\SyliusBrandPlugin\Entity\BrandImage">
+    <class name="Loevgaard\SyliusBrandPlugin\Model\BrandImage">
         <property name="file">
             <constraint name="Image">
                 <option name="maxSize">10M</option>
