--- tests/Application/Entity/Product.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Loevgaard\SyliusBrandPlugin\Application\Entity;
-
-use Loevgaard\SyliusBrandPlugin\Entity\ProductInterface as LoevgaardSyliusBrandPluginProductInterface;
-use Loevgaard\SyliusBrandPlugin\Entity\ProductTrait;
-use Sylius\Component\Core\Model\Product as BaseProduct;
-
-class Product extends BaseProduct implements LoevgaardSyliusBrandPluginProductInterface
-{
-    use ProductTrait;
-}
