+++ tests/Application/Model/Product.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Loevgaard\SyliusBrandPlugin\Application\Model;
+
+use Loevgaard\SyliusBrandPlugin\Model\ProductInterface as LoevgaardSyliusBrandPluginProductInterface;
+use Loevgaard\SyliusBrandPlugin\Model\ProductTrait;
+use Sylius\Component\Core\Model\Product as BaseProduct;
+
+class Product extends BaseProduct implements LoevgaardSyliusBrandPluginProductInterface
+{
+    use ProductTrait;
+}
