+++ tests/Application/config/doctrine/model/Product.orm.xml
@@ -1,14 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<!--suppress XmlDefaultAttributeValue -->
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                                       http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <mapped-superclass name="Loevgaard\SyliusBrandPlugin\Entity\BrandImage" table="loevgaard_brand_image">
-        <many-to-one field="owner" target-entity="Loevgaard\SyliusBrandPlugin\Entity\Brand" inversed-by="images">
-            <join-column name="owner_id" referenced-column-name="id" nullable="false" on-delete="CASCADE" />
+    <mapped-superclass name="Tests\Loevgaard\SyliusBrandPlugin\Application\Model\Product" table="sylius_product">
+        <many-to-one field="brand" target-entity="Loevgaard\SyliusBrandPlugin\Model\Brand" inversed-by="products">
+            <join-column name="brand_id" on-delete="SET NULL" />
         </many-to-one>
     </mapped-superclass>
