+++ tests/Application/config/services.yaml
@@ -11,5 +11,5 @@ sylius_product:
     resources:
         product:
             classes:
-                model: Tests\Loevgaard\SyliusBrandPlugin\Application\Entity\Product
+                model: Tests\Loevgaard\SyliusBrandPlugin\Application\Model\Product
                 repository: Tests\Loevgaard\SyliusBrandPlugin\Application\Doctrine\ORM\ProductRepository
