+++ tests/Behat/Context/Setup/BrandContext.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Tests\Loevgaard\SyliusBrandPlugin\Behat\Context\Setup;
 
 use Behat\Behat\Context\Context;
-use Loevgaard\SyliusBrandPlugin\Entity\BrandInterface;
+use Loevgaard\SyliusBrandPlugin\Model\BrandInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
