+++ tests/DataFixtures/ORM/resources/products.yml
@@ -30,7 +30,7 @@ Sylius\Component\Core\Model\ProductTranslation:
         description: <paragraph(2)>
         translatable: "@product_setono_sticker"
 
-Tests\Loevgaard\SyliusBrandPlugin\Application\Entity\Product:
+Tests\Loevgaard\SyliusBrandPlugin\Application\Model\Product:
     product_setono_mug:
         brand: "@brand_setono"
         fallbackLocale: "en_US"
